Old win32 binaries have been deleted.

+++ generator/fasthash.cc
@@ -66,5 +66,5 @@ uint64_t fasthash64(const void *buf, size_t len, uint64_t seed) {
 
 uint32_t fasthash32(const void *buf, size_t len, uint32_t seed) {
   uint64_t h = fasthash64(buf, len, seed);
-  return h - (h >> 32);
+  return (uint32_t)(h - (h >> 32));
 }
